backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-jose
passlib[argon2]
pydantic
pytest
httpx
python-multipart